requirement.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
mysqlclient==2.2.4
numpy==1.26.2
platformdirs==4.1.0
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.25.0
